irc: try to treat all IRC messages as UTF-8
Incoming non-ASCII strings do not initially have the Perl-internal utf8 flag
set, which causes the bytes to be interpreted as the default encoding
(typically, Latin-1).  Normally we're firing those bytes back to IRC so it
doesn't matter, but for ~define, the YAML serialization applies a second layer
of UTF-8 encoding, causing the factoids to become gibberish when Destult is
reloaded.

Testing done:
Using the following additional command, verified that "isutf8 á" replies "Is".
  ISUTF8: |-
    sub {
      use utf8;
      my( $kernel, $heap, $who, $what, $src, $dest, $replypath ) = @_;
      if( utf8::is_utf8( $what ) ) {
        $kernel->post( $src, $replypath, "Is", $dest);
      } else {
        $kernel->post( $src, $replypath, "Is not", $dest);
      }
    }

Verified that "define mu as µ" now puts "MU: µ" in factoids.yaml.  Previously,
it would be "MU: Âµ".

Verified with Wireshark that 'ß' is sent to IRC as the bytes C3 9F, rather than
DF (its Latin-1 value, and an invalid UTF-8 sequence).